Move pgclientSetup
