+++ src/steps.js
@@ -3,7 +3,7 @@
 
 const fs = require('fs');
 
-const { execShell, exec, execStdOut } = require('./exec');
+const { exec, execStdOut } = require('./exec');
 const learningContent = require('./replicate-learning-content');
 const enrichment = require('./enrichment');
 const logger = require('./logger');
@@ -16,24 +16,6 @@ const REPLICATION_MODE = {
   INCREMENTAL: 'incremental',
   TO_EXCLUDE: 'none',
 };
-
-// dbclient-fetch assumes $HOME/bin is in the PATH
-async function setupPath() {
-  await execShell('mkdir -p "$HOME/bin"');
-  // eslint-disable-next-line no-process-env
-  process.env.PATH = process.env.HOME + '/bin' + ':' + process.env.PATH;
-}
-
-function installPostgresClient(configuration) {
-  return exec('dbclient-fetcher', [ 'pgsql', configuration.PG_CLIENT_VERSION ]);
-}
-
-async function pgclientSetup(configuration) {
-  if (process.env.NODE_ENV === 'production') {
-    await setupPath();
-    await installPostgresClient(configuration);
-  }
-}
 
 async function dropCurrentObjects(configuration) {
   // TODO: pass DATABASE_URL by argument
@@ -198,7 +180,6 @@ module.exports = {
   fullReplicationAndEnrichment,
   getTablesWithReplicationModes,
   importLearningContent,
-  pgclientSetup,
   restoreBackup,
   REPLICATION_MODE,
 };
